requirements.txt
requests>=2.31.0
python-dotenv
fastapi
uvicorn
qdrant-client
sentence-transformers
torch
slowapi
